requirements.txt
streamlit
openai
pandas
streamlit-drawable-canvas
aisuite[all]
gspread
google.oauth
google.auth